+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/RuntimeContext.kt
@@ -79,7 +79,7 @@ class RuntimeContext(
           jsiContext.installJSIForBridgeless(
             this,
             jsRuntimePointer,
-            reactContext.runtimeExecutor!!
+            reactContext.catalystInstance.runtimeExecutor!!
           )
         } else {
           jsiContext.installJSI(