+++ tools/src/check-packages/checkDependenciesAsync.ts
@@ -46,6 +46,7 @@ const IGNORED_PACKAGES = [
   'expo-store-review', // package: expo-constants
   'expo-system-ui', // package: react-native-web
   'expo-updates', // cli: @expo/plist, debug, getenv - utils: @expo/cli, @expo/metro-config, metro
+  'jest-expo', // package: expect, jest-snapshot, react, server-only
 ];
 
 /**