+++ packages/expo-gl/CHANGELOG.md
@@ -13,6 +13,10 @@
 
 ### 💡 Others
 
+### ⚠️ Notices
+
+- Added support for React Native 0.75.x. ([#30034](https://github.com/expo/expo/pull/30034) by [@gabrieldonadel](https://github.com/gabrieldonadel))
+
 ## 14.0.2 — 2024-04-24
 
 ### 🐛 Bug fixes