+++ packages/jest-expo/CHANGELOG.md
@@ -20,6 +20,10 @@
 
 - Mock files from the `src` folder instead of `build`. ([#29702](https://github.com/expo/expo/pull/29702) by [@tsapeta](https://github.com/tsapeta))
 
+### ⚠️ Notices
+
+- Added support for React Native 0.75.x. ([#30034](https://github.com/expo/expo/pull/30034) by [@gabrieldonadel](https://github.com/gabrieldonadel))
+
 ## 51.0.2 — 2024-05-16
 
 _This version does not introduce any user-facing changes._